+++ database/migrations/2026_04_13_010644_create_product_reviews_table.php
@@ -18,9 +18,9 @@ return new class () extends Migration {
             $table->foreign('customer_id')->references('id')->on('customers')->nullOnDelete();
             $table->unsignedInteger('order_id')->nullable()->index();
             $table->foreign('order_id')->references('id')->on('orders')->nullOnDelete();
-            $table->unsignedSmallInteger('master_product_id')->nullable()->index();
+            $table->unsignedBigInteger('master_product_id')->nullable()->index();
             $table->foreign('master_product_id')->references('id')->on('master_assets')->nullOnDelete();
-            $table->unsignedSmallInteger('product_id')->nullable()->index();
+            $table->unsignedBigInteger('product_id')->nullable()->index();
             $table->foreign('product_id')->references('id')->on('products')->nullOnDelete();
             $table->unsignedTinyInteger('rating_main')->index();
             $table->unsignedTinyInteger('rating_a')->index()->nullable();
